Add files via upload
I have written a program that accepts as input a text file where each line describes a set of cards. This program determines whether each set is or is not valid.

A card has three attributes: a shape (square or circle or triangle), a colour (blue or red or green) and a pattern (spot or solid or stripe).
Two cards are distinct if they do not have the same shape or colour or pattern i.e. none of their attributes match.
Two cards are the same if they have the same shape and colour and pattern i.e. all their attributes match.
A set of three cards is valid if it consists of (i) three distinct cards, or (ii) three of the same card.