Add files via upload
I wrote a java program that calculates the point of intersection of two lines that 2 different ships travel on and then determines which of two points (ships)  is closest to the intersection. Each line has an equation in the form aX + bY + c =0. The program reads input from a file and in the file each line is in the form of 'a b c currentXpoint currentYpoint '
The program uses two pairs of lines to calculate their point of intersection and which one is closer to the intersection .